Address review feedback and fix branch redirect hash crash
- Extract ProjectTitleEditor to web-common for reuse across surfaces
- Extract DisabledCloudFeatures component; de-dupe explore/canvas lists
- Convert Exit link to Button type="ghost" for consistency
- Hoist getPreviewUrl IIFE to module-level helper
- Rename previewUrl -> editPreviewUrl in BranchesSection
- Drop url.hash from project layout redirect (unavailable in load)
